refactor: separate reward estimation so it can be tested in isolation
