+++ src/utils/index.ts
@@ -5,3 +5,4 @@ export * from "./constants";
 export * from "./bitcoin-core-client";
 export * from "./issueRedeem";
 export * from "./storage";
+export * from "./rewards";
